Implement export all/selected based on extension
The plugin works with metadata specified in the layer name of the form
```
LayerName e=png s=2 p=path m=10
```
Where
- e: extension
- s: scale
- p: custom path
- m: 10

This commit implements only `e=...` and the supported formats are `png`,
`jpg` & `jpeg`.

The plugin will create a tree folder structure in the current directory,
based on the Krita tree document structure, starting with a folder named
`export`. See the following illustrations:

The 2 implemented operations are:
- batch export all layers that are exportable (that is, that have the
  `e=...` metadata info)
- export selected layers even if no metadata is supplied. If no `e=...`
  info is given, then the `png` extension will be used

The plugin tries to be smart about the folder structure when exporting.
It will attempt to create only necessary folder, that is, only when a
group layer has any children or grandchildren that are exportable
themselves. The folder names themselves are stripped of metadata
information and in all cases, whitespaces are reduced to one `_`.