Export finally works, but only for RGBA & 8-bit dept
After a long series of bug-tracking and understanding how stuff works in
Krita Python API we finally have export of of individual layers with
different scales and support for .png & .jpg.

There are two main problems among other things (like having indicated
which functions are concurent and which aren't) with the Krita Python API
implementation:

1. Node.setColorProfile & Document.setColorProfile seem to be bugged so
they're unusable. This means that without a long conversion table, and
even with that, is impossible to convert all color modes Krita supports
to Pillow images. So beacuse of this reason I'm restricting the plugin
to export only RGBA 8-bit depth color profiles. There's a status bar
indicating success or failure on export.

2. Using file layers with .kra files isn't supported, that's why the
filtering/option to select .kra file types in the file layer dialog are
turned off. Krita can't properly render other .kra files so at the
moment it's impossible to support exporting from those layers. This is
odd because if we manually set the file layer to open a .kra file, it
properly loads it. It's just that the pixel data retrieval fails. In any
case, this is a deep Krita internal problem that needs to be fixed at
some point.